# Reduce allocations using nil cursors and literal value cursors