* Replaces individual literal cursors with a single interface value. This has the benefit that the value is immediately wrapped in an `iface`, removing `convT2I` calls and subsequent allocations every time the value is read.
* Statically allocates the nil value cursors, saving further allocations as they are stateless and safe for concurrent use.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkIntegerIterator_Next-8     82.8          22.7          -72.58%

benchmark                           old allocs     new allocs     delta
BenchmarkIntegerIterator_Next-8     3              0              -100.00%

benchmark                           old bytes     new bytes     delta
BenchmarkIntegerIterator_Next-8     32            0             -100.00%
```

###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass