@jwilder 2017-07-31
@jsternberg You are more familiar with this code than I am. LGTM though.