@stuartcarnie 2017-07-31
@jsternberg / @jwilder what should we put in the CHANGELOG for this one?