@jsternberg 2017-07-31
I'm not really sure. Likely goes in the features section and references this PR. Just a one sentence blurb about optimizing the memory usage for some cursors works. Enough for a person who is perusing the changelog to click the link for more information if they desire.